Time: 23 ms (24.27%) | Memory: 18.3 MB (55.44%) - LeetSync

+++ 2163-kth-distinct-string-in-an-array/kth-distinct-string-in-an-array.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     string kthDistinct(vector<string>& arr, int k) {
     map<string, int> b;
-    long n = arr.size();
+    int n = arr.size();
     for (int i = 0; i < n; i++) {
         b[arr[i]]++;
     }
